GhostPDF - don't free a buffer on error, it has been freed
OSS-fuzz #55443 with eps2write

The font has been corrupted and throws an error, which frees the buffer
we created and passed in. But we then try and free the buffer again
leading to a crash (on Windows at least).

Update the comment to note that ownership is transferred regardless of
success, and remove the code freeing the buffer.