Add "imagematrices_are_untrustworthy" to gs_data_image_t
The pdfwrite device can 'rewrite' the image matrixes in a
particularly nasty way that is not easy to avoid. It does this
to be able to capture images with masks of a different size.

Unfortunately, this breaks any possibility of us using the
imagematrix to figure out which source region of the image
we actually need for clipping.

Introduce a new flag to capture when this happens so we can
avoid relying on an untrustworthy imagematrix.