GhostPDF - fix type 3 font 'd0' arguments
Bug #706400 "Type 3 CharProcs with wy == 0 argument to d0 not drawn"

The width array argument was being initialised incorrectly due to
copy/paste errors. We were copying the second argument to both array
elements. If the second argument was 0 (which the PDF Reference states
it must be) then we would end up drawing a glyph with 0 width and height
and so would not draw it at all.

Fix that, and get the elements in the correct order.

This shows some tiny (1 pixel) alignment shifts in type 3 characters in
a couple of files.