GhostPDF - Fix visibility of some OCMDs
Bug #706399 "Optional content not drawn when OCMD has empty OCGs"

The file uses the BDC operator and the named optional content references
an Optional Content Management Dictionary rather than an Optional
Content Group.

The OCMD has an OCGs member whose value is an empty array. Because it
also has no 'P' entry we were defaulting to the 'AnyOn' value. Then we
checked each of the (non-existent) OCGs to see if any are 'on'. Since
none are, we determined the visibility to be 'off'.

However, according to the PDF 1.7 specification, p366, table 4.49, the
OCGs entry, if the OCGs are empty (or not present) then the OCMD should
'have no effect'. Its not obvious what that means, exactly, so I've
chosen to interpret it as 'should use the OCProperties->D->BaseState as
the visibility state'.

The BaseState is optional and so (as in this case) may not be present
in which case we use the default value of 'on'.