oss-fuzz 55443: Dangling reference in extreme out-of-memory condition
If we've exhausted memory to the extent that we cannot even run the Postscript
error handler properly, we can end up with references to the arrays that
contain the record of the stacks' content in $error lasting until the final
garbage collection, with the arrays themselves having been freed by the
end-of-job restore.

So explicitly remove those references and clear the operand stack before we
tackle the final stages of shutting down.