pdfwrite - preserve both true and false stroke adjustment
Bug #706407 "pdfwrite does not write /SA false unless it has previously been set to true"

The pdfwrite device only writes graphics state parameters which have
changed. The default value for the stroke adjust was set to false, so if
an input file set it to false, then pdfwrite would not write the value
to the output file.

The value is set to false because that is what the PDF Reference states
the default to be. However, it appears that Acrobat defaults to true.
This leads to different rendering for files which have /SA false and
files which do not set /SA. Interestingly at low resolution Ghostscript
also defaults to true.

The only way to deal with this is to have the initial value in the
viewer graphics state be neither true nor false. Then if an input file
never sets SA we will not write it, but if it does set it, no matter
what it gets set to, we will write it to the output.

This change results in a number of differences when cluster tested at
low resolution, and leads to a number of output files being reported as
increasing in size, naturally since we now write SA values when we would
not have before.

Checking the diffs we see that the files now match the rendering of the
original input file, whereas the file output by pdfwrite did not match
the original file, because the /SA false had been dropped.