Bug 706380: Fix argument quote handling following hash.
When parsing arguments, we have special handling for quote
marks. We interpret them as literal quotes, unless they are
either at the start of an argument, or immediately following
an '=' sign, so that:

 ThisIsAFileWithA"InTheName.pdf

and

 -sOutputFile="File 1.pdf"

both work as you'd expect.

Unfortunately, in the dim, dark, uncharted past that is
Ghostscript history, someone decided that it'd be a spiffy
idea for people to be able to use the hash character (that
one that I can't type in here because git will think it's
a comment character) as a replacement for '='.

The special handling to keep track if we had just seen an
equals sign was not coping with having just seen a hash
mark. Fixed here.

We also take the opportunity to fix a memory leak in the
handling of @files - ensuring we free the allocated block
for the stream when we return from reading from it.