Bug 706409: Fix infinite loop in shading code.
In the new subdivide_patch_fill_floats code (and indeed
in subdivide_patch_fill), there was a flaw in the logic.

Consider a rectangle (0,0) (100,100) and a patch with
v[0].p = v[1].p = v[2],p = v[3].p = (0,101).

We would therefore arrive at check_rot_bottom, and calculate
m0 = m1 = 101.

This is not <= rect.q.y, so we would continue through the
calculation, setting changed without actually changing the
coords of the patch.

The fix is to test against the same value in the middle of the
loop as we do at the end.

Do this throughout the code.