Ghostscript - Improve the mswinpr2 device's handling of printer names
Bug #706353 " mswinpr2 device fails on printer names longer than 31 characters"

This commit was supplied by Bernd Watermann, so I'm not completely
competent to write the commit log. I believe the main chnages are
to use wide character versions of the API calls (and variables)
for better support of non-English printer names and because the
old version had a 31-character name limit.

I've tested this with the old bugs referenced in the report, a
regular cluster regression run (even though it does not test
this device, and by manufacturing and checking a number of
virtual printers with names intended to explore the limits.