pdfwrite - permit fonts to be converted to linework based on the name
No bug report for this one, the patch was submitted by Justin after a
conversation on Discord (#ghostscript, 11th February 2023).

The goal was to extend the 'NoOutputFonts' control, which turns all
text in a PDF file into linework, such that the decision could be
made based on the font name.

The supplied patch does so by adding two new controls, AlwaysOutline
and NeverOutline, which can only be set from PostScript (because
the font names are in an array). Text using fonts appearing in the
AlwaysOutline array will be converted to linework regardless of the
setting of NoOutputFonts. If NoOutputFonts is set to true, then text
using fonts appearing in the NeverOutline array will still use the
font, and will not be converted to linework.