GhostPDF - process a null object for the page of a /Dest
We were not allowing the page of a /Dest to be null, which means
'do not change the current value'.

Add handling to cater for that, and set the destination page to 0,
which pdfwrite interprets as meaning it should emit a 'null'.

At the same time, check that the /Dest array contains a name object in
element 2. We may need to add specific testing of the name in future to
ensure it is valid.