Tweaks to pattern scaling.
Spun out of bug 706373, though not believed to be a fix for it.

When we calculate pattern matrixes and bboxes etc, floating point
accuracy conspires to work against us, in particular where we
end up mapping boxes back and forth via inverse matrices.

The rounding is such that we can end up expanding the number of
repeats that we have to render.

We try to be smarter about that (or at least to give saner bboxes
and matrices that are easier to debug) by employing some smart
("sane") rounding.