Epson ESC device - add fuzzy matching for media sizes
Bug #706367 "ESC/Page driver does not set page size correctly"

The PLRM indicates (Table 6.2, page 401, PageSize) that page dimensions
are considered to match if they are within a tolerance of 5 units in
each direction. This device (like others) was insisting on a precise
match.

Just add the tolerance.

I have no way to test this, other than looking at the content of the
output from the device, written to file, which appears to be correct.