GhostPDF - Dest names need not be C strings
Test file supplied in confidence by customer 820

The file has an Outlines entry where the /Dest has a name which is
in UTF16-BE. The code was dealing with strings as C strings, which
fails here because UTF16-BE text can contain NULLs.

In this commit; alter the comparison to use memcmp instead of strncmp
and pass the length of the PDF string as a parameter so that we do not
need to use strlen.