pdfwrite - detect duplicate images with SMask entries
Bug #706418 "Detect duplicate images with SMask"

We were not adding the SMask entry to the image dictionary until after
we had checked to see if the image (stream and dictionary) was identical
to a prior instance. The prior instance, of course, was written and
hashed with the SMask entry present, so the two instances never matched.

Fix by adding the SMask entry before we complete the image and check for
duplicates.

This exposed a bug in smask_image_check. The pointer to a character
being tested was never incremented, leading to an infinite loop, Clearly
this code was never previously being used.