Bug 706257: CIDFont glyph ordering issues with pdfi/pdfwrite
This exposed a couple of issues:

Firstly, and most importantly, when pdfwrite uses the callback to retrieve the
glyph index for text in a CIDFont, it uses the descendant font, not the Type 0,
as I originally thought. For embedded CIDFonts, that didn't cause a problem, but
for substituted CIDFonts it meant the glyph decoding callback did not have
access to the decoding table.

Secondly, fixing that exposed some byte ordering issues, where Unicode codes
read from the ToUnicode CMap differed in byte order from codes read from the
decoding table.