ps2write - fix emission of stroke adjust in type 3 fonts
No bug, but the specimen file /tests_private/comparefiles/bug688967.ps

This turned up in weekly testing. The file output by ps2write was
producing errors. This (eventually) turned out to be because a type 3
font contained a CharProc which was using an ExtGState Resource which
was not in the /Resources dictionary of the Font.

Why this happens is not entirely clear to me, it doesn't with pdfwrite.
What seems to be happening is that we are only capturing the Resources
required for the font when we run the BuildChar. In this case we have
multiple instances of type 3 fonts which are using the same BuildChar
but only the first apparently has the full set of Resources.

Since we don't need to write stroke adjustment as an ExtGState for
PostScript, and PDF output seems to work correctly, just do it
differently for PostScript output and write the stroke adjustment
directly.

This is a fix for commit 95db9c543abdeb391d852aad7f0c5c419f808069 which
I hadn't realised seems to have actually added the ability to write out
stroke adjustment at all.