Bug 706351: Fix display device with planar separations.
Using the display device with separations in chunky mode, we
insist on 8 components all the time.

In planar mode, the calculations go wrong, and we get crashes.
Update it here so that we don't crash.

Planar separations will use a maximum of 64 components.

In the case of a PDF file where the number of components in use
are known up front, we will run with the correct number of
components.

In the case of a PS file where the number of components is unknown
until rendering has completed, we run with the maximum number of
components, and the display device user can just take the required
number at the end.