Fix memory leak seen in apitest using gpdl.
A C parameter list leaks when used in gpdl. Presumably this is
not seen in gs operation because of garbage collection.

We already have a function to allocate a C parameter list
(gs_c_param_list_alloc). We've always been releasing this
(gs_c_param_list_release), but that just releases the contents
not the list itself. Add a new function (gs_c_param_list_free)
that both releases and frees, and call this as appropriate.