Graphics library - drop all copied components of a glyph on error
Bug #705476 "Some Russian letters disappear in PDF"

The problem here is quite subtle. We have multiple embedded TrueType
fonts which all have the same /FontName and so pdfwrite attempts to
'merge' the fonts into one, because it thinks they are the same font.

When the glyph cannot be embedded into the type 42 font we punt and
drop back to creating a type 3 font.

However.... Some of the glyphs in this file are compound glyphs, and the
first component can be added to the Type 42 font, but the second one
cannot. So we throw an error and don't add the second glyph, but we
leave the initial glyph in place. The error causes us to fall back to
type 3 for this glyph.

Then we encounter the same compound glyph again. Checking the font we
see that it is already embedded, and we don't check the components
because we assume that if we embedded the glyph we also embedded any
component glyphs. So we go ahead and use the embedded glyph and that
then fails to produce the correct outline.

When adding a component glyph to a font fails we need to 'unadd' the
initial glyph and any other components, so that we don't end up trying
to use any of them inappropriately.

For almost all fonts this is relatively straight-forward, but for
TrueType (type 42) fonts it is more complicated, so I've added another
member to the copied fonts procs so that we can have a specific method
for each font type.

So far the only case I've found is the TrueType font in this bug report
and since this has been present in the code forever it is clearly an
extremely rare case. Although I've added the code for the other font
types I'm not really able to test these terribly well.