Bug 706389: Fix versioning format consistency
The main code issue with this bug was that the banner printed on startup is
printed from Postscript, and Postscript's cvs operator doesn't allow for
dictating the number of digits it outputs so the number 00 will always end up as
the string "0", or 01 as "1". So our 10.01.0 version would be printed as
"10.1.0".

To address this, as a ".revisionstring" entry to systemdict, created during
startup, the string is created in C, so we control the format.

The remaining issues need to be address as part of the release process.