Tweak parameter testing to use an 'argis' macro.
At various places in the code, we test for parameters from a
param_list by using:

 if (!strncmp(param, "ParamValue", strlen("ParamValue"))

This is bad, because as well as matching "ParamValue" and
"ParamValue=", it will also match "ParamValueSomething".

Also, at various places in the code, we don't call strlen
(understandably, cos that's a runtime function call to retrieve
a constant value), and just wire in the constant value. But
in at least 1 location, we've got the constant value wrong.

Accordingly, move to using an 'argis' macro that tests correctly
and calculates the length at compile time.