gdev_prn_size_buf_planar: Initialise mdev.
We create a device structure on the stack, and fill in just
enough of it to be able to use existing code to return the
size.

Unfortunately, this produces valgrind warnings as some of
the proc setup code trips over the device procs not all being
zerod to start with.

For the sake of valgrind, just clear the device when we
create it. This is not something that happens often, so
the speed hit should be negligible.