Fix size_t -> int overflow in cmd_get_buffer_space.
When asking for the free space available, we have to be careful
when returning what might be a size_t in an int. In particular
this causes problems with 3G bufferspaces resulting in errors
as the value they return becomes a negative int.

The fix here is to clip the available buffer space at INT_MAX.

This is fine, but reworking the code to use an explicit
size_t return would probably be better eventually.