pdfwrite: Don't free copied font object when it's in use
If we spot early on that that we can't create a complete copy of a font (i.e.
if the font is incrememntally defined), we discard the attempted complete copy
and set the "complete" pointer to the same as the "copied" (subsetted) pointer.

But it turns out that, later, there is also a code path that requires discarding
the "complete" font, which wasn't accounting for the possibility that the two
pointers pointed to the same font object.

So, we now check whether the two pointers are the same, and if so, we simply
null the "complete" pointer.

Fixes oddly hard to reproduce segfaults from the nightly and weekly tests.