pdfwrite - Fix missing shading case
Bug #706453 "Element missing while converting pdf from rgb to cmyk using Ghostscript"

When converting a shading to an image (for ps2write or certain cases of
colour converting shadings with pdfwrite) it is possible that the
shading functions only call fill_trapezoid() and not fill_path().

This can result in us reaching dump_converted_image with
cvd->mask_is_empty and cvd->path_is_empty both true and we would then not
write out the rendered image.

Because of the convoluted nature of this code (which I am convinced is
sub-optimal!) simply setting either mask_is_empty and/or path_is_empty
to false results in incorrect rendering. So here we add a new boolean
'filled_trap' and add a case to deal with that after testing all the
other conditions.

This results in distinct progressions with ps2write and thee files:
/tests_private/comparefiles/Bug689189.pdf
/tests_private/comparefiles/Clarke-Tate-Manns-Chinese.ai

The file /tests_private/comparefiles/Bug691775.pdf shows diffs that
are imperceptible by eye.