Bug 706462: Fix heap overflow bug in command line dict parsing.
In the command line dictionary parsing, we were assuming
that there would be a close dict marker (">>") in the string
buffer after parsing completes. When that wasn't there
we might overrun the buffer.

Now check for it being there and raise a syntax error if not.

Thanks to Younseok Choi for the bug report.