Bug 706463: Fix buffer overflow seen in bjc600.
This code is poorly documented, and an unholy mix of "int", "word"
and "ulong" pointers, but I think I've understood the problem
enough to solve it.

The bug uses a single pixel wide device. It's therefore dealing
with lines that are expected to be 3 bytes wide. It allocates
buffers based on this number.

But it actually processes a larger number of pixels per line -
namely the number of pixels that will fit into the expected
amount of storage, rounded up to a 'word'. When words are 64bits
this makes much more difference than when they are 32.

Accordingly, we allow for this extra when allocating our buffers.

This should be a harmless change, as effectively all I've done is
allocated a buffer to be slightly larger than before.

Thanks to Youngseok Choi for the report.