Pdfwrite - don't count clipped text for auto rotation
Bug #706443 " AutoRotatePages=/PageByPage can be incorrect when there is text outside the TrimBox"

The bug file has text which lies off the page and so is entirely clipped
out and not passed through to the output file. However, because we count
text in pdf_text_begin, before we determine that the text is clipped,
it is still counted for the purposes of auto-rotation.

In this commit; add a new boolean in the pdf text enumerator to track
whether the text has been entirely clipped out. Initialise the flag in
gdev_pdf_text_begin. Move the counting for the purposes of auto-rotation
from pdf_text_begin to pdf_text_release so that we have a chance to
determine if the text is clipped before counting it. In
pdf_process_string, when we determine the text is clipped out, set the
flag.

I also fixed an annoying compiler warning at the same time, in
gdev_pdf_text_begin.

This causes changes in :

tests_private/comparefiles/541_623.pdf
tests_private/comparefiles/541.pdf
tests_private/comparefiles/Altona_Technical_v20_x4.pdf
tests_private/comparefiles/Clarke-Tate-Manns-Chinese.ai
tests_private/ps/ps3cet/13-29.PS

Because text that is clipped out is no longer considered, which
changes the orientation of these pages. One is a progression, one is
debatable and three are arguably regressions.

So I do not intend to commit this until after the release currently
underway is complete (10.1.0). This will give us some time to see
how it performs internally.