GhostPDF - Use the correct encryption configuration for streams
Bug 706474 "Segfault in gs for PDF with encrypted streams and unencrypted strings"

This is a simple typo, unencrypted streams are handled by an early exit
from pdfi_filter, so we should never see a CRYPT_IDENTITY case, just
as the comment says. But we were looking at the string encryption,
not the stream encryption. A one letter typo....