Bug 696884: Handle subsampled components correctly for JPEG2000 images.
Previously if a PDF had an JPEG2000 image using subsampling which was
not a multiple of the image width, rendering would appear staggered.
The upsampling code didn't take the up to dx - 1 upsampled samples
into account at the end of the scanline.

The JPEG2000 image in the PDF in the bug is 32 by 32 pixels where component 0
is not subsampled (dx/dy are both 1) and components 1 and 2 are subsampled
(dx/dy are both 2).

Thanks for the maths in this commit Robin!