GhostPDF - fix another category of broken PDF files
Bug #706500 "regression: new PDF engine can't handle a PDF file"

In this case the PDF file has been edited, and the original xref table
is invalid:

xref
1 7
0000000000 65535 f
0000000009 00000 n

That should, of course, be "0 7". The fact that it is '1' means we read
the expected offsets off by one each time. So object 0 gets read as
object 1 (which is then defined as a free object. Object 1 gets read
as object 2, so object 2 has the offset which points to object 1, etc.

So when we try to get the first page, object 4, we actually read object
3. Nothing is wrong apart from the fact that it's the wrong object but
obviously we check what object we got, and return an error because it
isn't the object we wanted.

This commit simply inserts another attempt to repair the file if
dereferencing an object results in no error, but no object was read
(can't see any way this is possible, but still) or the object we read
was a token not a PDF object, or the object number didn't match the
expected number (this case).