pdfwrite - do not reverse the order of clip subpaths
Bug #706523 "Error during merge Pdf Function : "*** C stack overflow. Quiting...""

The example file uses (on page 10) an enormously complicated clip path.
I suspect this was a chroma-keyed image or something similar which has
been downgraded by pdfwrite (which produced the original file) to a
fill and a complex clip.

The code for writing clips in pdfwrite uses a recursive function in
order to reverse the order of the subpaths. None of us can see a sane
reason for reversing the subpaths, so this commit simply does away with
that, which means we don't need the recursion and don't fall foul of
highly complex clip paths. (Note on Linux this did not fail, presumably
because the C exec stack is larger. But a more complex example
presumably would fail)

A few PXL files, when converted to PDF and the PDF rendered, show
dropouts at low resolution (65 dpi) but these files are already dropping
most of the strokes anyway at this resolution so I'm not going to worry
about the small difference. A few other files exhibit single pixel
differences at higher resolution, but again nothing of concern.