Bug 706530: Tweak shading optimisation code.
We've had several bugs in this code to do with getting stuck in
infinite loops, due to the accuracy of calculating midpoints.

Here, we tweak the tests so that they always choose NOT to loop
if the midpoint is close enough to the edge, so we should
always be safe.

Also, move the "completely out" clipping test to the top of the
loop where it can be more effective.