Ghostscript - convert the system default paper size to lower case
Bug 706544 "Unknown .defaultpapersize: (Letter). error shown at startup"

The problem appears to be that the system libpaper is configured to
have a default paper size of 'Letter' rather than the all lower case
'letter', so Ghostscript doesn't have a matching media size.

libpaper says that lower case is 'preferred' but there's apparently no
reason why we couldn't also have (for example) A4 instead of a4.

We can tackle this by making the system defined paper size lower case
before returning it to Ghostscript. However, a few of the paper sizes
in statusdict do have upper case characters in their name, so we need
to duplicate those few sizes as lower case. The contents of statusdict,
and in particular the defined media sizes, are not specified so we're
OK to do this.