GhostPDF - prevent buffer overrun when evaluating functions
OSS-fuzz bug #57745

The problem in the report is that the BlackGeneration function is a 1-in
3-out function. It is required to be a 1-in, 1-out function. The result
was that the evaluation was writing 3 floats to a 1 float buffer.

Check the parameters of the function to make sure it is of the correct
size before trying to evaluate it.

I also desk-checked all the other uses of functions; most were already
checking the function parameters but I found two more cases which were
not. Fix the /Separation and DeviceN tint transform so that we check the
number of inputs and outputs to make sure they are correct.