pdfwrite - don't replace UTF16 > 0x800 with the replacement glyph
Bug #706551 "ps2pdf corrupts Unicode title in PDF 1.4 XML metadata"

In fact the XMP metadata string is not 'corrupted' we simply replace all
UTF16 values > 0x800 with the Unicode replacement glyph. This was to do
with some investigations back in 2016 which never came to anything.

Simply promote UTF16 values > 0x800 to 3 bytes.