Bug 706552: more tweaks of Type 1 encodings
Infuriatingly, there are at least a couple of fonts that a PDF interpreter just
has to know are symbolic, regardless of the flag in their font descriptor.

Wingdings and ZapfDingbats and their variants.

Worse, contrary to the spec, if a font object for one of those fonts includes an
encoding entry which is a name object, we have to ignore it.