pdfwrite - support surrogate pairs in XMP UTF-8 strings
Bug #706551 "ps2pdf corrupts Unicode title in PDF 1.4 XML metadata"

This is not in fact part of the original report and should have been reported
separately. Still....

The bug arose from mer removing a partial implementation of surrogate
pairs when we started using Coverity, because Coverity complained about
the code and it was simpler to remove it. I clearly forgto to go back
and finish it.

This just adds code to deal with the (documented as unusual) case of
UTF-16 surrogate pairs. Seems to work with all the tests I can concoct.